Shifting Tkinter to python2 bindings,crash fix, modules folder, separate interfaces for recorder and visualiser
