+++ .gitignore
@@ -1,6 +1,7 @@
 .vscode/
 .pylintrc
 lsiv_output/
+output/
 .idea/
 __pycache__/
 *.pyc
