+++ modules/stitch.py
