+++ output/2020-04-19 18-16-27/info.json
@@ -1,0 +1,87 @@
+{
+    "File_Name": "26-16-IIDC.svs",
+    "Level_Details": [
+        {
+            "Width": 1,
+            "Height": 1,
+            "Level": 0
+        },
+        {
+            "Width": 2,
+            "Height": 2,
+            "Level": 1
+        },
+        {
+            "Width": 4,
+            "Height": 4,
+            "Level": 2
+        },
+        {
+            "Width": 7,
+            "Height": 7,
+            "Level": 3
+        },
+        {
+            "Width": 13,
+            "Height": 14,
+            "Level": 4
+        },
+        {
+            "Width": 25,
+            "Height": 27,
+            "Level": 5
+        },
+        {
+            "Width": 50,
+            "Height": 53,
+            "Level": 6
+        },
+        {
+            "Width": 99,
+            "Height": 105,
+            "Level": 7
+        },
+        {
+            "Width": 197,
+            "Height": 209,
+            "Level": 8
+        },
+        {
+            "Width": 394,
+            "Height": 418,
+            "Level": 9
+        },
+        {
+            "Width": 788,
+            "Height": 835,
+            "Level": 10
+        },
+        {
+            "Width": 1575,
+            "Height": 1670,
+            "Level": 11
+        },
+        {
+            "Width": 3150,
+            "Height": 3339,
+            "Level": 12
+        },
+        {
+            "Width": 6300,
+            "Height": 6678,
+            "Level": 13
+        },
+        {
+            "Width": 12600,
+            "Height": 13356,
+            "Level": 14
+        },
+        {
+            "Width": 25199,
+            "Height": 26711,
+            "Level": 15
+        }
+    ],
+    "File_Path": "D:/26-16-IIDC.svs",
+    "Level_Count": 16
+}
