--- requirements.txt
@@ -1,12 +1,0 @@
-entrypoints==0.3
-lxml==4.4.1
-mccabe==0.6.1
-numpy==1.17.3
-opencv-python==4.1.1.26
-openslide-python==1.1.1
-pandas==0.25.2
-Pillow==6.2.1
-pkg-resources==0.0.0
-python-dateutil==2.8.0
-pytz==2019.3
-six==1.12.0
